refactor(raijin): align owner boundaries
