--- yarn/raijin/src/filesystem/files/discover.interfaces.ts
@@ -1,10 +1,0 @@
-import type { PortablePath } from '@yarnpkg/fslib'
-
-export interface FileDiscoveryOptions {
-  readonly cwd: PortablePath
-  readonly patterns: ReadonlyArray<string>
-  readonly ignore?: ReadonlyArray<string>
-  readonly dot?: boolean
-}
-
-export type FileDiscovery = (options: FileDiscoveryOptions) => Promise<Array<PortablePath>>
