--- yarn/raijin/src/filesystem/files/discover.ts
@@ -1,5 +1,0 @@
-import type { FileDiscovery } from './discover.interfaces.js'
-
-import { discover }           from './adapters/fast-glob/discover.js'
-
-export const discoverFiles: FileDiscovery = discover
