+++ yarn/raijin/src/filesystem/files/discovery/discover.ts
@@ -1,0 +1,5 @@
+import type { FileDiscovery } from './discover.interfaces.js'
+
+import { discover }           from './adapters/fast-glob/discover.js'
+
+export const discoverFiles: FileDiscovery = discover
