+++ yarn/raijin/src/filesystem/index.ts
@@ -1,6 +1,6 @@
-export type { FileDiscovery }        from './files/discover.interfaces.js'
-export type { FileDiscoveryOptions } from './files/discover.interfaces.js'
+export type { FileDiscovery }        from './files/discovery/discover.interfaces.js'
+export type { FileDiscoveryOptions } from './files/discovery/discover.interfaces.js'
 
-export { discoverFiles }             from './files/discover.js'
+export { discoverFiles }             from './files/discovery/discover.js'
 export { toNativePath }              from './paths.js'
 export { toPortablePath }            from './paths.js'
