--- yarn/raijin/src/runtime/prettier/plugin/getters/index.ts
@@ -1,1 +1,0 @@
-export * from './prettier-plugin.getter.js'
