--- yarn/raijin/src/runtime/prettier/plugin/getters/printers.getter.ts
@@ -1,5 +1,0 @@
-import type { GetPrintersReturn } from '../interfaces/index.js'
-
-import { printers }               from '../printers/index.js'
-
-export const getPrinters = async (): GetPrintersReturn => printers
