+++ yarn/raijin/src/runtime/prettier/plugin/import-sort/index.ts
@@ -1,4 +1,4 @@
-export * from './import-sort-typescript.parser.js'
-export * from './import-sort.preprocess.js'
-export * from './import-sort.api.js'
-export * from './import-sort.style.js'
+export * from './typescript.parser.js'
+export * from './preprocess.js'
+export * from './matchers.js'
+export * from './style.js'
