--- yarn/raijin/src/runtime/prettier/plugin/interfaces/index.ts
@@ -1,1 +1,0 @@
-export type * from './prettier-plugin.interfaces.js'
