--- yarn/raijin/src/runtime/prettier/plugin/interfaces/prettier-plugin.interfaces.ts
@@ -1,17 +1,0 @@
-import type { Printer } from 'prettier'
-import type { Parser }  from 'prettier'
-
-export type PrintersType = Record<string, Printer>
-export type ParsersType = Record<string, Parser>
-
-export interface GetPrettierPluginOptions {
-  workspacePackageNames?: ReadonlyArray<string>
-}
-
-export type GetPrettierPluginReturn = Promise<{
-  printers: PrintersType
-  parsers: ParsersType
-}>
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type GetPrintersReturn = Promise<Record<string, any>>
